Fixes context.Canceled error
The issue was not about artemis properly killed / exited. It was more
of the dead go routine reading the in-memory pipe. Using a
ContextCommand introduced a series of other bugs as an interrupted
artemis exiting with success, would now report a context.Canceled. To
avoid to whitelist this error, it now simply use the signal of an
interrupt to ask for a clean termination.